+++ threatbus/test_data.py
@@ -14,6 +14,7 @@ from data import (
     SnapshotRequestDecoder,
     SnapshotEnvelopeEncoder,
     SnapshotEnvelopeDecoder,
+    ThreatBusSTIX2Constants,
 )
 
 
@@ -33,6 +34,7 @@ class TestJsonConversions(unittest.TestCase):
             modified=self.created,
         )
 
+        self.sighting_source = "VAST"
         self.sighting_id = "sighting--df0f9a0e-c3b6-4f53-b0cf-5e9c454ee0cc"
         self.sighting_context = {
             "ts": "2017-03-03T23:56:09.652643840",
@@ -63,7 +65,6 @@ class TestJsonConversions(unittest.TestCase):
             },
             "event_type": "alert",
             "_extra": {"vast-ioc": "172.31.129.17"},
-            "source": "VAST",
         }
         self.sighting = Sighting(
             id=self.sighting_id,
@@ -71,8 +72,9 @@ class TestJsonConversions(unittest.TestCase):
             modified=self.created,
             sighting_of_ref=self.indicator_id,
             custom_properties={
-                "x_threatbus_sighting_context": self.sighting_context,
-                "x_threatbus_indicator": self.indicator,
+                ThreatBusSTIX2Constants.X_THREATBUS_SIGHTING_CONTEXT.value: self.sighting_context,
+                ThreatBusSTIX2Constants.X_THREATBUS_INDICATOR.value: self.indicator,
+                ThreatBusSTIX2Constants.X_THREATBUS_SOURCE.value: self.sighting_source,
             },
         )
 
@@ -98,11 +100,11 @@ class TestJsonConversions(unittest.TestCase):
         self.assertEqual(py_dict["snapshot_type"], MessageType.SIGHTING.value)
         self.assertEqual(py_dict["snapshot_id"], self.snapshot_id)
         self.assertEqual(py_dict["snapshot"], self.snapshot.total_seconds())
-        self.assertEqual(py_dict["type"], type(SnapshotRequest).__name__.lower())
+        self.assertEqual(py_dict["type"], SnapshotRequest.__name__.lower())
 
     def test_valid_snapshot_request_decoding(self):
         encoded = f"""{{
-            "type": "{type(SnapshotRequest).__name__.lower()}",
+            "type": "{SnapshotRequest.__name__.lower()}",
             "snapshot_type": {MessageType.SIGHTING.value},
             "snapshot_id": "{self.snapshot_id}",
             "snapshot": {self.snapshot.total_seconds()}
@@ -132,7 +134,7 @@ class TestJsonConversions(unittest.TestCase):
         self.assertEqual(py_dict["body"]["id"], self.indicator_id)
         self.assertEqual(py_dict["body"]["pattern"], self.pattern)
         self.assertEqual(py_dict["body"]["pattern_type"], self.pattern_type)
-        self.assertEqual(py_dict["type"], type(SnapshotEnvelope).__name__.lower())
+        self.assertEqual(py_dict["type"], SnapshotEnvelope.__name__.lower())
 
         py_dict = json.loads(encoded_envelope_sighting)
         self.assertEqual(py_dict["snapshot_type"], MessageType.SIGHTING.value)
@@ -140,13 +142,14 @@ class TestJsonConversions(unittest.TestCase):
         self.assertEqual(py_dict["body"]["created"], format_datetime(self.created))
         self.assertEqual(py_dict["body"]["sighting_of_ref"], self.indicator_id)
         self.assertEqual(
-            py_dict["body"]["x_threatbus_sighting_context"], self.sighting_context
+            py_dict["body"][ThreatBusSTIX2Constants.X_THREATBUS_SIGHTING_CONTEXT.value],
+            self.sighting_context,
         )
-        self.assertEqual(py_dict["type"], type(SnapshotEnvelope).__name__.lower())
+        self.assertEqual(py_dict["type"], SnapshotEnvelope.__name__.lower())
 
     def test_valid_snapshot_envelope_decoding(self):
         encoded_envelope_indicator = f"""{{
-            "type": "{type(SnapshotEnvelope).__name__.lower()}",
+            "type": "{SnapshotEnvelope.__name__.lower()}",
             "snapshot_type": {MessageType.INDICATOR.value},
             "snapshot_id": "{self.snapshot_id}",
             "body": {{
@@ -164,7 +167,7 @@ class TestJsonConversions(unittest.TestCase):
         self.assertEqual(read_back, self.snapshot_envelope_indicator)
 
         encoded_envelope_sighting = f"""{{
-            "type": "{type(SnapshotEnvelope).__name__.lower()}",
+            "type": "{SnapshotEnvelope.__name__.lower()}",
             "snapshot_type": {MessageType.SIGHTING.value},
             "snapshot_id": "{self.snapshot_id}",
             "body": {{
@@ -174,8 +177,9 @@ class TestJsonConversions(unittest.TestCase):
                 "type": "sighting",
                 "spec_version": "2.1",
                 "id": "{self.sighting_id}",
-                "x_threatbus_sighting_context": {json.dumps(self.sighting_context)},
-                "x_threatbus_indicator": {self.indicator.serialize()}
+                "{ThreatBusSTIX2Constants.X_THREATBUS_SIGHTING_CONTEXT.value}": {json.dumps(self.sighting_context)},
+                "{ThreatBusSTIX2Constants.X_THREATBUS_INDICATOR.value}": {self.indicator.serialize()},
+                "{ThreatBusSTIX2Constants.X_THREATBUS_SOURCE.value}": "{self.sighting_source}"
             }}
         }}"""
         read_back = json.loads(encoded_envelope_sighting, cls=SnapshotEnvelopeDecoder)